requirements.txt
django>=5.0
pytube>=10.12.0
moviepy>=1.0.3
ffmpeg-python>=0.2.0
pytubefix>=6.1.1